Removed unessesary old forms of Integer type: contrib dir

+++ contrib/groovy.grailsproject/src/org/netbeans/modules/groovy/grailsproject/plugins/GrailsPluginSupport.java
@@ -364,7 +364,7 @@ public class GrailsPluginSupport {
                 try {
                     try {
                         Integer retValue = future.get();
-                        if (retValue.intValue() != 0) {
+                        if (retValue != 0) {
                             broken = true;
                         }
                     } catch (InterruptedException ex) {
